requirements.txt
mockserver-friendly-client==0.4.1
requests==2.25.1
